Extract worker spawn logic into launch module
